+++ src/main/java/matt/pass/mojaryba/web/SearchController.java
@@ -1,6 +1,6 @@
 package matt.pass.mojaryba.web;
 
-import matt.pass.mojaryba.domain.fish.FishService;
+import matt.pass.mojaryba.domain.fish.FishSearchAndTopService;
 import matt.pass.mojaryba.domain.fish.dto.FishDto;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,18 +11,20 @@ import java.util.List;
 
 @Controller
 public class SearchController {
-    private FishService fishService;
 
-    public SearchController(FishService fishService) {
-        this.fishService = fishService;
+    private final FishSearchAndTopService fishSearchAndTopService;
+
+    public SearchController(FishSearchAndTopService fishSearchAndTopService) {
+        this.fishSearchAndTopService = fishSearchAndTopService;
     }
+
     @GetMapping("/search-form")
     String searchForm(){
         return "search-form";
     }
     @GetMapping("/search")
     String searchFishes(Model model, @RequestParam String search) {
-        final List<FishDto> foundFishes = fishService.searchFishes(search);
+        final List<FishDto> foundFishes = fishSearchAndTopService.searchFishes(search);
         model.addAttribute("heading", "Znalezione okazy");
         model.addAttribute("description",
                 "Poniżej znalezione okazy. Szukaliśmy po tytule oraz opisach okazów.");
