+++ src/main/java/matt/pass/mojaryba/web/admin/AdminPanelController.java
@@ -1,6 +1,7 @@
 package matt.pass.mojaryba.web.admin;
 
 import matt.pass.mojaryba.domain.user.User;
+import matt.pass.mojaryba.domain.user.UserAdminService;
 import matt.pass.mojaryba.domain.user.UserService;
 import matt.pass.mojaryba.domain.user.dto.UserAdministrationDto;
 import org.springframework.stereotype.Controller;
@@ -16,10 +17,12 @@ import java.util.List;
 @Controller
 public class AdminPanelController {
 
-    private UserService userService;
+    private final UserService userService;
+    private final UserAdminService userAdminService;
 
-    public AdminPanelController(UserService userService) {
+    public AdminPanelController(UserService userService, UserAdminService userAdminService) {
         this.userService = userService;
+        this.userAdminService = userAdminService;
     }
 
     @GetMapping("/admin/uzytkownicy")
@@ -35,13 +38,13 @@ public class AdminPanelController {
     }
     @GetMapping("/admin/usun-uzytkownika/{id}")
     String deleteUser(@PathVariable long id, RedirectAttributes redirectAttributes){
-        userService.deleteUserById(id);
+        userAdminService.deleteUserById(id);
         sendNotyfication(redirectAttributes, "Użytkownik został usunięty");
         return "redirect:/admin/uzytkownicy";
     }
     @GetMapping("/admin/uzytkownik/{id}")
     String userEditPage(Model model, @PathVariable long id) {
-        final UserAdministrationDto user = userService.findAdministrationUserById(id);
+        final UserAdministrationDto user = userAdminService.findAdministrationUserById(id);
         model.addAttribute("user", user);
         return "admin-panel-user-edit";
     }
@@ -50,7 +53,7 @@ public class AdminPanelController {
         if (userService.chechExistByNick(nick)) {
             sendNotyfication(redirectAttributes, "Podany nick jest już zajęty");
         }else {
-            userService.adminEditUserNick(nick, id);
+            userAdminService.adminEditUserNick(nick, id);
           sendNotyfication(redirectAttributes, "Nick został zmieniony");
         }
         return "redirect:/admin/uzytkownik/" + id;
@@ -62,14 +65,14 @@ public class AdminPanelController {
         if (userService.chechExistByEmail(email)) {
             sendNotyfication(redirectAttributes, "Podany email jest już zajęty");
         } else {
-            userService.adminEditUserEmail(email, id);
+            userAdminService.adminEditUserEmail(email, id);
            sendNotyfication(redirectAttributes, "Adres email został zmieniony");
         }
         return "redirect:/admin/uzytkownik/" + id;
     }
     @PostMapping("/admin/edytuj-haslo/{id}")
     String editUserPass(@PathVariable long id, @RequestParam String password, RedirectAttributes redirectAttributes){
-        userService.adminEditUserPass(password, id);
+        userAdminService.adminEditUserPass(password, id);
         sendNotyfication(redirectAttributes, "Hasło zostało zmienione");
         return "redirect:/admin/uzytkownik/" + id;
     }
@@ -79,25 +82,25 @@ public class AdminPanelController {
     }
     @GetMapping("/admin/dezaktywuj/{id}")
         String deactivateUser(@PathVariable long id, RedirectAttributes redirectAttributes){
-        userService.deactivateOrActivateUser(id);
+        userAdminService.deactivateOrActivateUser(id);
         sendNotyfication(redirectAttributes, "Konto zostało dezaktywowane");
         return "redirect:/admin/uzytkownik/" + id;
     }
     @GetMapping("/admin/aktywuj/{id}")
         String activateUser(@PathVariable long id, RedirectAttributes redirectAttributes){
-            userService.deactivateOrActivateUser(id);
+            userAdminService.deactivateOrActivateUser(id);
             sendNotyfication(redirectAttributes, "Konto zostało aktywowane");
             return "redirect:/admin/uzytkownik/" + id;
     }
-    @GetMapping("/admin/block/{id}")
+    @GetMapping("/admin/blokuj/{id}")
         String blockUser(@PathVariable long id, RedirectAttributes redirectAttributes) {
-        userService.blockUser(id);
+        userAdminService.blockUser(id);
         sendNotyfication(redirectAttributes, "Użytkownik został zablokowany");
         return "redirect:/admin/uzytkownik/" + id;
     }
-    @GetMapping("/admin/unblock/{id}")
+    @GetMapping("/admin/odblokuj/{id}")
         String unblockUser(@PathVariable long id, RedirectAttributes redirectAttributes) {
-        userService.unblockUser(id);
+        userAdminService.unblockUser(id);
         sendNotyfication(redirectAttributes, "Użytkownik został odblokowany");
         return "redirect:/admin/uzytkownik/" + id;
     }
