+++ src/main/java/matt/pass/mojaryba/web/admin/RegisterController.java
@@ -1,6 +1,7 @@
 package matt.pass.mojaryba.web.admin;
 
 import jakarta.validation.Valid;
+import matt.pass.mojaryba.domain.user.UserAdminService;
 import matt.pass.mojaryba.domain.user.UserService;
 import matt.pass.mojaryba.domain.user.dto.UserRegisterDto;
 import org.apache.commons.mail.EmailException;
@@ -12,10 +13,12 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 public class RegisterController {
-    private UserService userService;
+    private final UserService userService;
+    private final UserAdminService userAdminService;
 
-    public RegisterController(UserService userService) {
+    public RegisterController(UserService userService, UserAdminService userAdminService) {
         this.userService = userService;
+        this.userAdminService = userAdminService;
     }
 
     @GetMapping("/rejestracja")
@@ -39,7 +42,7 @@ public class RegisterController {
             } catch (EmailException e) {
                 System.err.println("Problem z wysyłką email");
                 e.printStackTrace();
-                userService.deleteUserByEmail(user.getEmail());
+                userAdminService.deleteUserByEmail(user.getEmail());
                 model.addAttribute("heading", "Błąd podczas rejestracji");
                 model.addAttribute("description",
                         "Nie udało się wysłać wiadomości z linkiem aktywacyjny. Dokonaj rejestracji ponownie");
